Ring, Circle and Line now accept float values to size parametres

+++ include/shapes/ring.h
@@ -3,12 +3,12 @@
 class Ring : public Circle
 {
 public:
-    Ring(size_t aPosY, size_t aPosX, float aR, size_t aThickness, uint32_t aColor);
+    Ring(size_t aPosY, size_t aPosX, float aR, float aThickness, uint32_t aColor);
     ~Ring();
 
-    void setThickness(size_t aThickness);
-    size_t getThickness() const;
+    void setThickness(float aThickness);
+    float getThickness() const;
 
 private:
-    size_t thickness;
+    float thickness;
 };
